+++ pkg/hostman/monitor/hmp.go
@@ -294,7 +294,7 @@ func (m *HmpMonitor) DeviceAdd(dev string, params map[string]interface{}, callba
 	m.Query(fmt.Sprintf("device_add %s,%s", dev, strings.Join(paramsKvs, ",")), callback)
 }
 
-func (m *HmpMonitor) MigrateSetDowntime(dtSec float32, callback StringCallback) {
+func (m *HmpMonitor) MigrateSetDowntime(dtSec float64, callback StringCallback) {
 	m.Query(fmt.Sprintf("migrate_set_downtime %f", dtSec), callback)
 }
